Modify member controller
