+++ routes/web.php
@@ -15,6 +15,4 @@ Route::get('/', 'PagesController@home');
 
 Route::get('about', 'PagesController@about');
 
-Route::get('members', 'MembersController@index');
-
-Route::get('members/{member}', 'MembersController@show');
+Route::resource('members', 'MemberController');
